Fix the subdir package import
Learned how to import the subdir package in golang project.
The rule is to use the module name written in the top of go.mod as the
root name.
For example, if the module name is "x/y" and the subdir package is "z",
import becomes `import "x/y/z"`.